- Resolves conflicts in docs/changelog.md, converters.py, masks.py,
  key_points/annotators.py, utils/conversion.py, utils/video.py
- Keeps OpenCV optional-dependency changes (try/except imports,
  ensure_cv2_installed) alongside develop's new features and fixes
- Adds from __future__ annotations and typing imports from develop
  to converters.py; adds CompactMask import to masks.py; adds logger
  to key_points/annotators.py and utils/video.py; adds TypeVar F to
  conversion.py
- Changes cv2 = None  # type: ignore to # type: ignore[assignment]
  (specific error code; avoids unused-ignore warning under strict mypy)
- Annotates CV2_FONT: int | None and adds assert before return to fix
  mypy no-any-return in annotators/core.py

---

+++ examples/speed_estimation/ultralytics_example.py
@@ -70,7 +70,7 @@ def main(
     )
     trace_annotator = sv.TraceAnnotator(
         thickness=thickness,
-        trace_length=video_info.fps * 2,
+        trace_length=int(video_info.fps * 2),
         position=sv.Position.BOTTOM_CENTER,
     )
 
@@ -79,7 +79,7 @@ def main(
     polygon_zone = sv.PolygonZone(polygon=SOURCE)
     view_transformer = ViewTransformer(source=SOURCE, target=TARGET)
 
-    coordinates = defaultdict(lambda: deque(maxlen=video_info.fps))
+    coordinates = defaultdict(lambda: deque(maxlen=int(video_info.fps)))
 
     with sv.VideoSink(target_video_path, video_info) as sink:
         for frame in frame_generator:
